fastapi/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.1.0
pydantic-settings>=2.0.0
pytest>=7.4.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
httpx>=0.24.1
databricks-sdk>=0.61.0
databricks-sql-connector==4.0.2
pandas>=2.0.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
google-auth==2.40.3
greenlet==3.2.3
h11==0.16.0
idna==3.10
pyasn1==0.6.1
pyasn1-modules==0.4.2
pydantic-core==2.33.2
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
sqlalchemy==2.0.41
sqlmodel==0.0.24
starlette==0.47.2
typing-extensions==4.14.0
typing-inspection==0.4.1
urllib3==2.5.0
uuid==1.30